pyproject.toml
[project]
name = "backend-smart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "openai>=2.32.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
    "xhtml2pdf>=0.2.17",
]
